CFY-1758 Refacoring
- 'utility' subdir was renamed to 'scripts'
- a new 'utils' module was created; context validation, api manager
  handling and several common functions were moved there
