+++ main/configure_minion.py
@@ -21,11 +21,12 @@ from cloudify.decorators import operation
 from cloudify.exceptions import NonRecoverableError
 from cloudify.exceptions import RecoverableError
 
-from validation import validate_context
+import utils
 
 
-_DEFAULT_MINION_ID_PATH = '/etc/salt/minion_id'
 _DEFAULT_MINION_CONFIG_PATH = '/etc/salt/minion'
+_DEFAULT_MINION_KEYS_DIR = '/etc/salt/pki/minion/'
+_DEFAULT_MINION_ID_PATH = '/etc/salt/minion_id'
 _YAML_LOADER = yaml.SafeLoader
 _YAML_DUMPER = yaml.SafeDumper
 
@@ -46,54 +47,77 @@ def _load_minion_config(path=_DEFAULT_MINION_CONFIG_PATH):
         )
         return {}
 
-def _write_to_protected_file(data, path):
-    p = subprocess.Popen(
-        ['sudo', 'tee', path],
-        stdin=subprocess.PIPE,
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT
-    )
-    output = p.communicate(input=data)
-    message = 'sudo tee {0} output:\n{1}'.format(path, output)
-
-    p.wait()
-    if p.returncode != 0:
-        ctx.logger.error(message)
-        raise RecoverableError(
-            'Failed to write to file {0}.'.format(path)
-        )
-    else:
-        ctx.logger.debug(message)
 
 def _save_minion_config(config, path=_DEFAULT_MINION_CONFIG_PATH):
-    ctx.logger.debug('Writing minion configuration to {0}'.format(path))
+    ctx.logger.info('Writing minion configuration to {0}'.format(path))
     data = yaml.dump(
         config,
         default_flow_style=False,
         Dumper=_YAML_DUMPER
     )
-    _write_to_protected_file(data, path)
+    utils.write_protected_file(data, path)
+
 
 def _save_minion_id(minion_id, path=_DEFAULT_MINION_ID_PATH):
-    ctx.logger.debug('Writing minion id to {0}'.format(path))
-    _write_to_protected_file(minion_id, path)
+    ctx.logger.info('Writing minion id to {0}'.format(path))
+    utils.write_protected_file(minion_id, path)
+
+
+def _authorize_minion(minion_id):
+    if _is_authorized(minion_id):
+        ctx.logger.info('Minion {0} is already authorized.'.format(minion_id))
+    else:
+        ctx.logger.info('Generating keys for minion {0}...'.format(minion_id))
+        private, public = _generate_key(minion_id)
+        try:
+            path = _DEFAULT_MINION_KEYS_DIR
+            subprocess.check_output(['sudo', 'mkdir', '-p', path])
+        except subprocess.CalledProcessError as e:
+            raise NonRecoverableError('Cannot create {0} directory: {1}\n'
+                                      .format(path, e.output))
+        utils.write_protected_file(private, path + 'minion.pem')
+        utils.write_protected_file(public, path + 'minion.pub')
+        ctx.logger.info('Minion {0} authorized.'.format(minion_id))
+
+
+def _is_authorized(minion_id):
+    mgr = utils.instantiate_manager()
+    resp, result = mgr.accepted_minions()
+    if resp.ok:
+        accepted = result['data']['return']['minions']
+        return minion_id in accepted
+    else:
+        ctx.logger.error('Got response {0}'.format(resp))
+        raise NonRecoverableError(
+            'Unable to get accepted minions from Salt API.'
+        )
+
+
+def _generate_key(minion_id):
+    mgr = utils.instantiate_manager()
+    resp, result = mgr.generate_accepted_key(minion_id)
+    if resp.ok:
+        ctx.logger.info('Generated key for minion {0}.'.format(minion_id))
+        keys = result['data']['return']
+        return keys['priv'], keys['pub']
+    else:
+        ctx.logger.error('Got response {0}'.format(resp))
+        raise NonRecoverableError(
+            'Unable to generate key for minion {0}.'.format(minion_id)
+        )
 
 
 @operation
 def run(*args, **kwargs):
-    validate_context(ctx.node.properties)
+    utils.validate_context()
+    minion_id = utils.get_minion_id()
+    subprocess.call(['sudo', 'service', 'salt-minion', 'stop'])
 
     ctx.logger.info('Updating minion configuration with blueprint data...')
-
-    subprocess.call(['sudo', 'service', 'salt-minion', 'stop'])
     config = _load_minion_config()
     config.update(ctx.node.properties['minion_config'])
     _save_minion_config(config)
-
-    minion_id = ctx.node.properties['minion_id']
-    if not minion_id:
-        minion_id = ctx.instance.id
     _save_minion_id(minion_id)
-    ctx.instance.runtime_properties['minion_id'] = minion_id
-
     ctx.logger.info('Minion configuration updated successfully.')
+
+    _authorize_minion(minion_id)
