+++ main/scripts/default_minion_installation.sh
